bug fixes, more tests, and some debug logging
The main issue was that drpc could deadlock in the case that
the manager received messages with no stream active. That's
fixed by having the select to send the message to the active
stream also attempt to enter the stream semaphore. If it can
then that means there's no stream, so it should just drop
the message.

I'm not convinced there are no more deadlocks. I probably need
to do some more low level tests to figure this out.

Change-Id: I56e1b2b12ca290743e7edecc4274b7bade7aa512