require id monotonicity across read calls
we require id monotonicity within a read call when reconstructing
a packet to make reading packets not require extra memory: we can
throw out any results when the id increases. but, we don't require
it across consecutive calls. Consider

    incomplete frame for packet 1
    complete frame for packet 2
    complete frame for packet 1

without this commit, that would successfully return packet 2, then
successfully return a partial result for packet 1, violating both
guarantees about complete packets but also monotonicity. this commit
adds an id that is stored across calls so that for the above case
we would return an error on the 3rd frame.

Change-Id: Ib979112a37b73b44e25009e2129d74b1bc8608c2