fixes to allow connection pooling
first, we add a `Closed` method to the `drpcconn.Conn` so
that you can do some best effort checking if the underlying
transport is still usable.

second, we ensure that the stream's context is only canceled
after we know all of the sends/receives for that stream are
completed. that way, users of the stream can be sure that
the conn that provided it can be reused.

Change-Id: I6814b8c6bda79d3e3a73672ef0d2d5374b06d84b