cancel should kill transport if necessary
if we're stuck in a write syscall and the context is canceled
the we must kill the transport in order to unblock. in order
to reuse the connection as much as possible especially given
that most contexts eventually become canceled, we don't kill
the transport if we're certain we have already sent a final
message for that stream to the remote end.

Change-Id: Ie06e426bc8266ca846232958b8ef742f82047854