cancel streams that are unmanaged
because clients terminate tcp connections if the stream is
unfinished when it is canceled, remotes should put their
stream into the cancel state when the tcp connection dies.
fortunately, the unmanaged case is exactly when the transport
dies as every other exit for those helper functions happens
when the stream is terminated for some other reason.

without this you end up with a lot of "stream closed by sending
error" messages on the servers after they attempted to send an
error to no one and then bubble that error up when they try to
close or whatever.

Change-Id: I12b02db5340eb2cc6613e5e8f396328bdb6b76c1