drpcmanager;drpcconn: add metadata into invoke message
in order to send tracing information across the wire, we need to extend
drpc invoke message to contain key/value pairs.
We will set the first two bytes of the message to be empty so we know
the message is using the new format and may contain meatadata.
Then we will proto buff to encode our invoke message so we can send it
across the wire.

Change-Id: I03a5af62c5b41e0855dd101faacd8d12c66c2cdb