stream/manager: fix cancellation propagation
in cases where a stream is not sending, like in a unitary
rpc after the invoke or a stream rpc that just isn't writing,
and a cancel happened, we would not close the transport even
though we would require more reads and writes to complete the
rpc because we considered the stream to be "finished."

change "finished" to mean that there will never be any more
reads or writes on the stream. this is accomplished by having
a read and write mutex and polling them to see if we can acquire
both after every read or write operation. if "terminated" is
set and we can even temporarily acquire both mutexes, because
terminated implies no new reads or writes will happen without
an error, we know there are no active reads or writes and no
traffic on the transport.

we used to check for finished while we were holding the write
mutex, but that caused problems when trying to acquire the read
mutex because of mutex ordering possibly leading to deadlocks. we
don't need to check all of the conditions atomically though, due
to "terminated" as explained above.

Change-Id: I0cd3237c33d4b1c999752e97b17c2bd5757d6fce