drpcmanager: wait for previous stream to finish before creating new
concurrent access to a drpc connection should be serialized by the
semaphore in the drpcmanager, but that semaphore is released when
the stream is terminated, not finished. that means that a previous
stream can still use the underlying transport even after the
manager allows a new stream to be created, causing issues.

the manager now will wait for the previous stream to be finished
before creating a new one. this doesn't cause any latency issues
because the transport is assumed to be reliable and stream-oriented
so if the close message is not being processed, then new messages
from the new stream would not be either.

Change-Id: Ie54b33023d9c5cb339fbd7ba9509abc6db581d8d