drpcmigrate subpkg and examples
This is the first step of moving two useful utilities out
of storj/storj - dialing with a header, and the listenmux.

We are interested in publishing a blog post with content about
drpc, and one of the main angles in the blog post is how others,
currently using grpc, can switch, so we want to make support
for switching a first-class part of the drpc project.

Change-Id: I5280841ba7861a7f94d5f9bdfe0bf32e24aac8d5