all: support different protobuf runtime libraries
This adds an Encoding type that describes how to read
and write the Message type. The generated code creates
an Encoding based on the protobuf library that was used
to generate the protobuf messages.

The default Encoding generated uses google.golang.org/protobuf
to marshal and unmarshal. This can be overridden with the
protolib flag during an invocation to protoc-gen-drpc as follows:

    protoc --drpc_out=protolib=github.com/gogo/protobuf:.

The argument is an import path. Both the google.golang.org/protobuf
and github.com/gogo/protobuf import paths are recognized by the
command to generate correct Encoding values. If an unrecognized
import path is passed, it is assumed that the import defines
functions that match the name and signature of the Encoding
interface, and it will generate a type that matches it.

Change-Id: I41e0c221e151ef2db8e8852d4ecb5ce9346f25af