internal/backcompat: add backwards compatability tests
this works by defining the following modules:

    ./backcompat
    ./backcompat/newservice
    ./backcompat/newservicedefs
    ./backcompat/oldservice
    ./backcompat/oldservicedefs
    ./backcompat/servicedefs

the ./backcompat module depends on ./backcompat/servicedefs
which just contains stub types/functions that match the
generated drpc code. that module then creates a Main function
that executes either as a server or a client, checking to
see that the generated rpcs work.

./backcompat/newservicedefs are generated every time go generate
runs and depend on the most recent drpc.

./backcompat/oldservicedefs are generated once with drpc version
v0.0.17 and are expected to never change again.

./backcompat/newservice uses some module replace tricks to replace
the ./backcompat module to the local directory, and the
./backcompat/servicedefs module to ./backcompat/newservicedefs. It
then calls the Main function, creating a server/client binary that
uses the ./backcompat/newservicedefs.

./backcompat/oldservice is the same way.

The tests in ./backcompat then execute both ./backcompat/newservice
and ./backcompat/oldservice in both client/server mode and point
them at each other.

Change-Id: I934e62fff3483d194e8a71d74cfbeb10a5a087b3