drpcmanager: make inactivity timeout off by default
the inactivity timeout breaks some expectations of users
who are dropping drpc in for grpc. many users expect
connections to keep working, and having the inactivity
timeout on by default means users will often get
unexpected errors. it's true that they should handle
the closed connection case for robustness reasons, but
perhaps we shouldn't force them to so frequently.
we should let them enable it if they want to, though.

i think we should consider making this on by default
once we have a high quality conn pool implementation in
this repo.

Change-Id: I73691e6236d0d1fdeaeb6e378ef75dd60759c764