drpcstream: remove three allocations
this removes an allocation for a context to return the
transport by passing it into the stream itself and
having the owned stream context value return it directly.

this removes another allocation by not having the manager
look at the Terminated signal. it does this by passing
in a channel that the stream sends once on. the manager
ensures that only one stream is active at a time, so
this is sufficient.

both of the above were accomplished by extending the
stream options with a set of internal options that can
only be modified and read with a new internal drpcopts
package.

this also removes an allocation by avoiding callbacks
by avoiding drpcwire.SplitN when writing. the compiler
was unable to prove that the passed in data slice did
not escape, so the string to []byte conversion when
passing in the rpc name had to allocate. with the
more inlined code, the compiler is now able to prove
it does not escape and avoids making a copy.

Change-Id: Id038f94eaeab32f31e6f88cbc06a502da63c9b05