drpcmanager: remove allocations introduced in go1.17
launching a goroutine ends up allocating a closure object
in go1.17 that it didn't before. this maybe has something
to do with the register abi?

in any case, because we only ever have 1 stream executing
at a time, we can have a single goroutine that just runs
the stream and pass it the information needed and launch
that single goroutine at the start. this doesn't have
an effect on any benchmarks, except for go1.17 benchmarks.

Change-Id: Ifadce4a73423c7e7c31c296a2f22fd7c11b674b2