drpcmanager: fix closing race condition
in some situations, newStream could be blocked forever
trying to send a stream into m.streams when the
manageStreams goroutine has already exited due to manager
terminating.

this changes the send into a select that can be canceled
by the manager terminating.

also, there was a race in the tests for the muxer where
rare failures could happen if the Route call was not
scheduled before the connection with the given prefix
was handled.

Change-Id: If2033d8b8e3b2c3e3a50e8a72337bc263458321e