cmd/protoc-gen-go-drpc: proto3 optional support
this additionally bumps the google.golang.org/protobuf
dependency to v1.27.1, and the nix flake dependencies
for protoc from v3.16.0 to v3.18.0.

adds an optional field into the grpccompat tests
rather than the integration tests because we generate
using gogoprotobuf sometimes in the integration tests
which does not support optional fields.

Change-Id: Idb52705b2e57588120a7f6af7bc2db36ddde3d09