drpcmanager: expose error from transport failures
A cancel can sometimes only be performed by closing the underlying
transport because that is the only way to interrupt a read/write
syscall. But, rather than assume that every error from the transport
failing is due to a cancel, only assume that an io.EOF means that.
This can help with debugging when the remote side dies due things
besides a "clean" close of the underlying transport.

Closes #31

Change-Id: Ia655d07984f6abe0492e7f9300d2638a1c82b4d4