drpcpool: add simple connection pool
this is a pool designed around the techniques used in the
connection pool in storj.io/common/rpc/rpcpool, but somewhat
simplified and generic in the key type. it adds about 10%
lines of code to the package, but this is a common issue
multiple users have hit and either given up or invented
themselves.

Change-Id: Ie86f523fb36283ac63b5757f6652a39971b1aca2