drpc{conn,manager,stream,pool}: soft cancel support
this commit adds "soft cancel" support where if a stream
is canceled, it attempts to send a small cancel message
instead of hard terminating the transport. the manager
will block creation of new streams until the stream has
sent this soft cancel message successfully, and terminate
if it can't send it right away. the manager exposes a
channel that is used by the pool to check if it is still
in a state where it is soft canceling so that the pool
can skip over connections in that state.

one thing to consider is that with connection pooling, a
soft cancel on a connection that has a bunch of data
in the send queue could still proceed and consider the
connection "unblocked" even though there will be some
head of line blocking for the next rpc. perhaps some
future code can either look at the send queue directly
or wait for acknowledgement from the remote side. the
latter option is not backwards compatible, though, so
some care would have to be taken to figure out how to
make that work.

this adds more tests to the pool and adds some fixes
found because of those tests. almost all of the tests
from the storj rpcpool cache have been ported except
for the fuzz test which is difficult to port due to
this design having the cache and pool combined.

Change-Id: I04a00b91e6a40fb3f1144a9a4f85bd2a9867d5f4