drpcstream: fix data race in packet buffer handling
the packet buffer would go into a finished state when
closed even if someone is still holding and using the
data slice from get, allowing put to unblock and causing
a data race. this adds a held boolean to the packet buffer
state so that close will wait for any handing of the
data buffer before going into the finished state.

fixes #48

Change-Id: I131fc7addb26153e31b68015e58a2e400c1ce8f5