all: some version bumps
also removes the check-errs lint because it wasn't providing
much value and was starting to become flaky with weird
compiler like errors (missing symbol resolution, etc). a
quick look at the source code of the package showed that it
was not much more than a wrapper around go analysis tools
so it's unclear what the problem is.

Change-Id: I34056cdca57b39aef41b2b311906e48cbd96afcd